move AkaoSnesModulation function definitions into a cpp file
